Extract advanced search into a component
