+++ app/assets/stylesheets/application.scss
@@ -6,6 +6,7 @@
 @import "icons";
 @import "mixins";
 @import "admin";
+@import "advanced_search";
 @import "layout";
 @import "participation";
 @import "milestones";
