+++ app/assets/stylesheets/layout.scss
@@ -8,7 +8,6 @@
 // 06. Forms
 // 07. Callout
 // 08. User account
-// 09. Search
 // 10. Official levels
 // 11. Tables
 // 12. Social
@@ -1477,40 +1476,6 @@ form {
   .time {
     font-size: $small-font-size;
     color: $text-medium;
-  }
-}
-
-// 09. Search
-// ----------
-
-.advanced-search {
-  float: left;
-  margin: $line-height 0;
-  position: inherit;
-
-  @include breakpoint(medium) {
-    float: right;
-    margin-bottom: 0;
-    margin-top: $line-height / 4;
-    position: absolute;
-    right: 0;
-  }
-}
-
-.advanced-search-form {
-
-  @include breakpoint(medium) {
-    > .column {
-      padding: 0;
-    }
-  }
-
-  select {
-    height: $line-height * 2;
-  }
-
-  .column.end.clear {
-    clear: both;
   }
 }
