+++ app/components/shared/advanced_search_component.html.erb
@@ -4,7 +4,7 @@
 
 <div class="row advanced-search-form">
   <%= form_tag request.path, id: "advanced_search_form", method: :get do %>
-    <div id="js-advanced-search" data-advanced-search-terms="<%= @advanced_search_terms.present? %>" style="display: none">
+    <div id="js-advanced-search" data-advanced-search-terms="<%= advanced_search.present? %>" style="display: none">
 
       <div class="small-12 column">
         <label for="search">
