+++ app/views/debates/index.html.erb
@@ -52,7 +52,7 @@
         </div>
       </div>
 
-      <%= render "shared/advanced_search" %>
+      <%= render Shared::AdvancedSearchComponent.new %>
 
       <%= render "shared/order_links", i18n_namespace: "debates.index" %>
